I updated scalar traits some.
One of the major changes I made was to add a configure-time test for <limits> and std::numeric_limits<>.  I then put in an #ifdef so that Teuchos::ScalarTraits can use this standard class for the definition of certain constants instead of the LAPACK routine.  Not only is this faster but my version of LAPACK that configure found did not have this routine.  When std::numeric_limits<> is available this should be used.

For now I only updated the eps() function but I can also update some of the other functions too.

Another change was to add support of generating NaN (the new nan() function) and a test for NAN or Inf (the new isnaninf() function).  I took the test for NaN/Inf from NOX and this seems to be a really good and portable test.  This test only requires that the hardward support the IEEE standard, not the C++ compiler (which has been a problem with checking x!=x as a test for NaN since most compilers are too smart and just return true at compile time).

I have run the test suite for both debug and optimized versions and they check out.