CcjSupport: I fixed some copyright headers and removed some unecessary comments/new lines.  I switched CcjLink from using System.err.println to using JpetraObject.println("ERR", msg).  I also cleaned up a confusing comment found in at least half of the Ccj Data object classes.
MatrixMarketIO:
I cleaned up the spacing and comments.  I added pattern read support for CisMatrix and MultiVector.  I also added integer read support to CisMatrix and MultiVector and made both throw a more explicit parse error when the type of matrix being read is unknown (probably due to an incorrect file format).  I fixed up the copyright headers where applicable as well.

Jpetra:
I made the CisMatrix/MultiVector serialized read/write methods only perform the read and write on the root node (node 0).  I've overridden clone() for both CisMatrix and MultiVector and made it so both objects can be cloned.  CisMatrix just copies the trees used to build the optimized form of the matrix and MultiVector copies the 2d array of doubles.  Both use the same vector space as the source instance.  I also overrode equals for both classes and for VectorSpace/ElementSpace in order to do equals on CisMatrix and MultiVector.  CisMatrix.multply works for A and A' for both row and column oriented matrices.