Changed BlockGmres so that it is configured by a parameter list, which is passed into the constructor.
Added small amount of code to BlockGmres to create a flexible variant, which can be
used if the parameter list has a parameter called "Variant" with the argument "Flexible".
So far this code checks out, but it is still being tested.  That is why it is in the belos_new
directory.  Flexible GMRES only accepts a right-preconditioner.

Added two tests for the Flexible GMRES, showing how to use it.  The tests use a static
preconditioner and thus the results are the same as right-preconditioned GMRES.  Initially
this was a sanity check, but a variable preconditioner (probably GMRES(5) will be put into
this example soon).