Addresses bug 944.
Ulrich Hetmaniuk discovered a performance bottleneck in Epetra_BasicDirectory
in the method GetDirectoryEntries where a linear search was being used on
a list that can potentially be very long.

The solution is to sort the list being searched and then use a binary or
bisection type search.

All epetra tests pass.