Initial version of a ParameterList interface for AztecOO. The new method is:   AztecOO::SetParameters(Teuchos::ParameterList& parameterlist); The parameterlist argument currently isn't const, probably should be. This method and associated implementation is conditionally compiled subject to the macro HAVE_AZTECOO_TEUCHOS, which is defined at configure time using the argument --enable-aztecoo-teuchos. If that argument isn't used, then AztecOO doesn't depend on Teuchos and doesn't have a ParameterList interface.
Doxygen comments in AztecOO.h give some explanation of what constitutes a
legal parameter name and value.

The AztecOO::SetParameters implementation makes use of mappings between
strings and the integer values that are used as offsets into Aztec's
options and params arrays. These mappings are in the file
AztecOO_string_maps.cpp, which is generated by the perl script
AztecOO_string_maps_builder.pl. This script is based on one that Kevin Long
has in TSFExtended. The script uses input from the file
AztecOO_string_maps.txt. This is currently pretty primitive. The goal is to
refine things to the point where AztecOO_string_maps.txt contains XML
from which documentation can be generated. Another goal is to have the
file az_aztec_defs.h (or portions thereof) also generated automatically
from the same input file so that there will be a single place to edit and
update parameters and macros for Aztec, and there won't be a risk of
az_aztec_defs.h getting out of sync with AztecOO_string_maps.cpp.