----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	src/Epetra_CrsMatrix.cpp
 ----------------------------------------------------------------------
"new"ed the creation of a temporary vector because it is not needed in
all cases.  We need to create it in all cases because we cannot conditionally
create it inside an if statement, but we can hold off on and dynamically
size the vector so that we don't use the space if we don't need it.

CrsMatrix tests pass.