Amesos_OverlappingSchwarzPreconditioner requires an Epetra_CrsMatrix in the constructor. Then, it creates the overlapping matrix, and subdivides the local graph in the required number of subdomains. Each subdomain can be solved using Amesos (locally), or LAPACK. As the inverse of each subdomain's matrix is created via Amesos_InverseFactory (given in input by the user), in principle any class that defined ApplyInverse() in an Epetra_Operator object can be used to solve the subdomain's problem.
Amesos_OverlappingSchwarzPreconditioner is an extension of
Amesos_OverlappingPreconditioner. The difference:

- the former requires an Epetra_CrsMatrix, can use an arbitrary overlap,
and on each processor the user can specify the number of
subdomains. This class can be used to define overlapping domain
decomposition preconditioners, with NumSubdomains > NumProcs.

- the latter requires an Epetra_RowMatrix, only minimal overlap can be
used, and one subdomain per processor.

Note however that Amesos_OverlappingSchwarzPreconditioner can more
memory demanding (as it must create the overlapping matrix).