----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	src/Epetra_CrsMatrix.cpp test/CrsMatrix/cxx_main.cpp
 ----------------------------------------------------------------------
Bug 966.

InvRowMaxs() and InvColMaxs() methods now work for matrices that have
wholy owned row or columns, respectively.  These row and columns can
be fully replicated on other processors, but at this time we are not
going to support matrices with partial sums on multiple processors.

Also, InvRowMaxs() will not work on a matrix that has rows that are owned
by mulitple processors when those rows are not fully replicated on all
processors that own at least one entry.  The analogous statemnent is
true for InvColMaxs().

I also wrote some simple tests for the methods.  These tests should be
improved at a later time.

All CrsMatrix tests pass.