Implement new VTK reader.
Improvements over the previous reader:
--------------------------------------

It is much more tolerante of extra and/or missing newlines.  The only
place it requires the presence of newlines is at locations where
binary data would start if the file where a binary file, as that is the
only place where the documetation I have for the file format explicitly
states that newlines are required.

Reports the line number in error messages for easier correction of
input files (or debugging if the reader.)

Can read all three of VTKs structured formats, if they specify 2D or
3D meshes.

Can read VTK attribute data into tags on mesh entities.


Known limitations:
------------------

Gives up if it encounters things it cannot read, such as unsuported
element types, field data, etc.  Perhaps it should have the option
of ignoring those entities and continuing to read the file.

Cannot read VTK "field" data.  The grouping of this data can't be
represented in tag data, and there's no way to represent global field
data, which isn't associated with mesh entities.

Cannot read sphere or edge elements, as Mesquite doesn't support those.

Will not read polygon data, as MeshImpl uses a fixed-length connectivity
list for elements, and thus cannot represent them.  The exception is if
polygons are encountered in a "POLYDATA" block (as opposed to an
"UNSTRUCTURED_GRID" block) and all the polygons are either 3- or 4-sided.

Will not read poly-vertex or poly-line data because Mesquite doesn't
support them (they aren't really mesh, but the could be mapped to
sphere elements and edges if Mesquite had those.)

Will not read triange-strips.  Two reasons: a) it is difficult to implement
given the representation used in MeshImpl for the mesh and b) it is unclear
whether or not the whole strip is to be treated as a single element for
the purpose of assigning attributes.  So I figured I'd wait until someone
actually had a file with these in that they wanted to read in Mesquite
before worring about implementing this.

Will not read 1D structured mesh blocks, because Mesquite doesn't do edges.

Doesn't handle higher-order elements.  VTK supports elements with mid-edge
nodes, but the fixed-length connectivity lists in MeshImpl don't have
sufficient space for them.  Also, it looked like it would break other
code if the number of nodes in the connectivity list for these elements
was more than the number for the linear form of the element.

Doesn't read binary files.  It wouldn't be that difficult to add, but
the VTK binary file format is horribly unportable, so unless someone
is actually using them, I'm not going to worry about implementing it.


Other changes:
--------------

I fixed up the tag storage in MeshImpl.  It no longer leaks the tag
storage and can store tag data for vertices.  I didn't make any
changes to the interface (which needs a lot of work) except to add
a method to delete tags.


Testing:
-------

This doesn't break any previously working tests.  I also tested
this code by writing some code that uses mesquite to read and
write back out a VTK file.  I then build to executables using that
code, one with these changes an one without.  For every file in
meshFiles/3D/VTK, I ran both executables on the file and compared
the output.

Added unit tests to test new capabilities.