Fix for Bug #1040.
When both "--with-nox-epetra-ml" and amesos (either directly via --enable-amesos
or indirectly) are active, libml.a needs libamesos.a to resolve
symbols at link time.  This caused build failures of some nox/test/epetra
problems.

The fix required using the macro, config/tac_arg_enable_can_use_package.m4,
to detect when the above scenario occurs and to then add libamesos.a to
the list of libraries used at link-time.

Unfortunately, this change touched all Makefile.in as well as a few other
configure files.