----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	test/AztecOO_SetParameters/cxx_main.cpp
 ----------------------------------------------------------------------
Usually we don't produce output when the verbose flag is not used, but
aztecoo tests do output ******* Test passed ****** so the lack of this
output in the AztecOO_SetParameters test has caused the test harness to
report a failure.  We will leave this output in there for a while as it
isn't harming anything at this time.