After noticing that bug 229 is still open (regarding a segv that was occurring in Epetra_MpiDistributor), I added a new function to epetra/test/Comm/cxx_main.cpp to make sure that the relevant code is covered. The bug appears to have been fixed long ago.
I made a small change in Epetra_MpiDistributor to correct an assignment
that wasn't real type-safe (assigning an int to a bool). Also added a
couple of trivial method implementations to Epetra_SerialDistributor so
it wouldn't produce an error if a couple of its methods are called by
the test I added to Comm/cxx_main.cpp.

I'll now close bug 229.